+++ apps/projects/views/team/delete_team_views.py
@@ -14,14 +14,33 @@
 #
 #   For permission inquiries, please contact: admin@Bimod.io.
 #
+
 import logging
 
 from django.contrib import messages
-from django.contrib.auth.mixins import LoginRequiredMixin
-from django.shortcuts import get_object_or_404, redirect
+
+from django.contrib.auth.mixins import (
+    LoginRequiredMixin
+)
+
+from django.shortcuts import (
+    get_object_or_404,
+    redirect
+)
+
 from django.views.generic import TemplateView
 
-from apps.projects.models import ProjectTeamItem
+from apps.core.user_permissions.permission_manager import (
+    UserPermissionManager
+)
+
+from apps.projects.models import (
+    ProjectTeamItem
+)
+
+from apps.user_permissions.utils import (
+    PermissionNames
+)
 from web_project import TemplateLayout
 
 logger = logging.getLogger(__name__)
@@ -31,21 +50,45 @@ class ProjectsView_TeamDelete(LoginRequiredMixin, TemplateView):
     def get_context_data(self, **kwargs):
         context = TemplateLayout.init(self, super().get_context_data(**kwargs))
         team_id = self.kwargs.get('pk')
-        team = get_object_or_404(ProjectTeamItem, pk=team_id)
+
+        team = get_object_or_404(
+            ProjectTeamItem,
+            pk=team_id
+        )
+
         context['team'] = team
+
         return context
 
     def post(self, request, *args, **kwargs):
         team_id = self.kwargs.get('pk')
-        team = get_object_or_404(ProjectTeamItem, pk=team_id)
+
+        ##############################
+        # PERMISSION CHECK FOR - DELETE_TEAMS
+        if not UserPermissionManager.is_authorized(
+            user=self.request.user,
+            operation=PermissionNames.DELETE_TEAMS
+        ):
+            messages.error(self.request, "You do not have permission to delete project teams.")
+            return redirect('projects:team_list')
+        ##############################
+
+        team = get_object_or_404(
+            ProjectTeamItem,
+            pk=team_id
+        )
+
         team_name = team.team_name
 
         try:
             team.delete()
+
         except Exception as e:
             messages.error(request, f'An error occurred while deleting the Team: {str(e)}')
+
             return redirect('projects:team_list')
 
         messages.success(request, f'Team "{team_name}" has been successfully deleted.')
         logger.info(f'Team deleted: {team_name}')
+
         return redirect('projects:team_list')
